CI: Remove sanitizers resource template
This template was added when the sanitizer builds needed more memory than all
of the other builds. We've since increased the amount of memory that the other
builds use beyond what's requested in this template, so it doesn't make sense
to keep it around.

(cherry picked from commit 4c6c89d8636b249041bdaabbaaaa7b83154295b6)