Order rule traversal in RuleMatcher matching operations by Rule index
This ordering fixes a test failure we're seeing on Alpine for the
signatures/tcp-end-of-match btest, since discrepancies in rule match traversal
could lead to discrepancies in corresponding event ordering. It looks safe to
rely on across platforms since the index is driven by signature load order,
which shouldn't deviate. If this somehow doesn't hold in the future, we'll only
wind up with a test failure, not incorrect match behavior.